pyproject.toml
[build-system]
requires = ["setuptools>=65.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "splice"
version = "1.0.0"
description = "SPLICE (Simple Platform for Locating Isoform Complexity Events): The definitive differential splicing analysis tool"
readme = "README.md"
license = { text = "MIT" }
authors = [
    { name = "Mariano Aguilar Vela", email = "mariano.aguilarvela@hdr.qut.edu.au" }
]
keywords = [
    "bioinformatics",
    "RNA-seq",
    "alternative splicing",
    "differential splicing",
    "transcriptomics",
    "splicing analysis"
]
classifiers = [
    "Development Status :: 4 - Beta",
    "Environment :: Console",
    "Intended Audience :: Science/Research",
    "License :: OSI Approved :: MIT License",
    "Natural Language :: English",
    "Operating System :: OS Independent",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Scientific/Engineering :: Bio-Informatics",
]
requires-python = ">=3.10"
dependencies = [
    "pysam>=0.22.0",
    "numpy>=1.24.0",
    "scipy>=1.11.0",
    "pandas>=2.0.0",
    "statsmodels>=0.14.0",
    "click>=8.1.0",
    "tqdm>=4.65.0",
    "pyarrow>=14.0.0",
    "pyfastx>=2.0.0",
    "biopython>=1.81",
    "scikit-learn>=1.3.0",
    "matplotlib>=3.7.0",
    "seaborn>=0.12.0",
    "numba>=0.58.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0",
    "pytest-cov>=4.0",
    "black>=23.0",
    "flake8>=6.0",
    "mypy>=1.0",
]
docs = [
    "sphinx>=5.0",
    "sphinx-rtd-theme>=1.0",
]

[project.urls]
"Homepage" = "https://github.com/mariano-aguilar-vela/splice"
"Bug Tracker" = "https://github.com/mariano-aguilar-vela/splice/issues"
"Documentation" = "https://splice.readthedocs.io"
"Source Code" = "https://github.com/mariano-aguilar-vela/splice"

[project.scripts]
splice = "splice.cli:main"

[tool.setuptools]
packages = ["splice"]

[tool.black]
line-length = 100
target-version = ['py310']
include = '\.pyi?$'
extend-exclude = '''
/(
  # directories
  \.eggs
  | \.git
  | \.hg
  | \.mypy_cache
  | \.tox
  | \.venv
  | build
  | dist
)/
'''

[tool.isort]
profile = "black"
line_length = 100
multi_line_mode = 3
include_trailing_comma = true

[tool.mypy]
python_version = "3.10"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = false
disallow_incomplete_defs = false
check_untyped_defs = true

[tool.pytest.ini_options]
minversion = "7.0"
addopts = "-ra -q --strict-markers"
testpaths = ["splice/tests"]
python_files = "test_*.py"
python_classes = "Test*"
python_functions = "test_*"
markers = [
    "integration: marks tests as integration tests",
    "benchmark: marks tests as benchmark tests",
    "slow: marks tests as slow (deselect with '-m \"not slow\"')",
]
